+++ spec/spec_helper.rb
@@ -9,9 +9,3 @@ SPEC_DIR = Pathname.new(File.dirname(__FILE__))
 Dir[SPEC_DIR.join("support", "**", "*.rb")].each {|f| require f}
 
 CONFIG = YAML.load_file(SPEC_DIR.join("config.yml")).with_indifferent_access
-
-RSpec.configure do |c|
-  c.before do
-    BloomRemitClient.sandbox = true
-  end
-end